feat: cleanup code, move entirely to go releaser
