--- .dockerignore
@@ -1,4 +1,0 @@
-node_modules
-bin
-config
-.github
