--- .goreleaser.yml
@@ -1,23 +1,0 @@
-# Documentation for this file: http://goreleaser.com
-before:
-  hooks:
-    - make dep
-builds:
-  - main: ./
-    id: &name minecraft-preempt
-    binary: *name
-    goos:
-      - linux
-      - darwin
-      - windows
-    goarch:
-      - amd64
-      - arm64
-    env:
-      - CGO_ENABLED=0
-archives: []
-checksum:
-  name_template: "checksums.txt"
-release:
-  # We handle releasing via semantic-release
-  disable: true
