--- .releaserc
@@ -1,15 +1,0 @@
-{
-  "plugins": [
-    "@semantic-release/commit-analyzer",
-    "@semantic-release/release-notes-generator",
-    ["@semantic-release/exec", {
-      "generateNotesCmd": "make release APP_VERSION=${nextRelease.version}"
-    }],
-    ["@semantic-release/github", {
-      "assets": ["dist/*.tar.gz", "dist/checksums.txt"]
-    }]
-  ],
-  "branches": [{
-    "name": "main"
-  }]
-}
