+++ .tool-versions
@@ -1,4 +1,3 @@
-golang 1.19.4
-nodejs 19.2.0
-goreleaser 1.13.1
-golangci-lint 1.50.1
+golang 1.20.1
+goreleaser 1.15.2
+golangci-lint 1.51.1
