+++ internal/cloud/docker/client.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/docker/docker/api/types"
 	dockerclient "github.com/docker/docker/client"
-	"github.com/jaredallard/minecraft-preempt/pkg/cloud"
+	"github.com/jaredallard/minecraft-preempt/internal/cloud"
 	"github.com/pkg/errors"
 )
